se crea propiedad bouncy_high y low para rebotar al tocar suelo y se corrige error GARRAFAL que definia props como byte y los usaba como int
git-svn-id: https://svn.code.sf.net/p/warcom/code/trunk/Warcom_Game_Engine@355 bdfe1ff2-85b1-43ac-a073-3d879cfd3ae1

+++ monsters.h
@@ -28,7 +28,7 @@ public
 	int   axisAlign     = 0;		//Alineacion del eje del grafico respecto caja colision
 	int   state 		= 0;   		//Estado de la entidad
 	int   prevState     = 0;		//Estado anterior
-	byte  props			= 0;		//Propiedades de la entidad
+	int  props			= 0;		//Propiedades de la entidad
 	struct colPoint[cNumColPoints] 	//Puntos deteccion colision de un objeto
 		int x;						//Offset X a sumar a la posicion del objeto
 		int y;						//Offset Y a sumar a la posicion del objeto
@@ -51,7 +51,7 @@ public
 	int   axisAlign     = 0;		//Alineacion del eje del grafico respecto caja colision
 	int   state 		= 0;   		//Estado de la entidad
 	int   prevState     = 0;		//Estado anterior
-	byte  props			= 0;		//Propiedades de la entidad
+	int  props			= 0;		//Propiedades de la entidad
 	struct colPoint[cNumColPoints] 	//Puntos deteccion colision de un objeto
 		int x;						//Offset X a sumar a la posicion del objeto
 		int y;						//Offset Y a sumar a la posicion del objeto
@@ -74,7 +74,7 @@ public
 	int   axisAlign     = 0;		//Alineacion del eje del grafico respecto caja colision
 	int   state 		= 0;   		//Estado de la entidad
 	int   prevState     = 0;		//Estado anterior
-	byte  props			= 0;		//Propiedades de la entidad
+	int  props			= 0;		//Propiedades de la entidad
 	struct colPoint[cNumColPoints] 	//Puntos deteccion colision de un objeto
 		int x;						//Offset X a sumar a la posicion del objeto
 		int y;						//Offset Y a sumar a la posicion del objeto
@@ -97,7 +97,7 @@ public
 	int   axisAlign     = 0;		//Alineacion del eje del grafico respecto caja colision
 	int   state 		= 0;   		//Estado de la entidad
 	int   prevState     = 0;		//Estado anterior
-	byte  props			= 0;		//Propiedades de la entidad
+	int  props			= 0;		//Propiedades de la entidad
 	struct colPoint[cNumColPoints] 	//Puntos deteccion colision de un objeto
 		int x;						//Offset X a sumar a la posicion del objeto
 		int y;						//Offset Y a sumar a la posicion del objeto
@@ -120,7 +120,7 @@ public
 	int   axisAlign     = 0;		//Alineacion del eje del grafico respecto caja colision
 	int   state 		= 0;   		//Estado de la entidad
 	int   prevState     = 0;		//Estado anterior
-	byte  props			= 0;		//Propiedades de la entidad
+	int  props			= 0;		//Propiedades de la entidad
 	struct colPoint[cNumColPoints] 	//Puntos deteccion colision de un objeto
 		int x;						//Offset X a sumar a la posicion del objeto
 		int y;						//Offset Y a sumar a la posicion del objeto
@@ -142,7 +142,7 @@ public
 	int   axisAlign     = 0;		//Alineacion del eje del grafico respecto caja colision
 	int   state 		= 0;   		//Estado de la entidad
 	int   prevState     = 0;		//Estado anterior
-	byte  props			= 0;		//Propiedades de la entidad
+	int  props			= 0;		//Propiedades de la entidad
 	struct colPoint[cNumColPoints] 	//Puntos deteccion colision de un objeto
 		int x;						//Offset X a sumar a la posicion del objeto
 		int y;						//Offset Y a sumar a la posicion del objeto
